Move commission related methods to Customer model
Allows us to avoid keeping track of the customer we are performing any commission operations on. We can implement a method within AddressBook to get a FilteredList of all commissions across all customers, but for customer specific commission operations, we can just put those under the Customer model.

Undo previous change from `UniqueCommissionList` to `Set<Commission>` in 85d2361

- The previous suggestion that led to the commit was made with the idea that we would need to change the customer's hash to trigger the UI rerender, but by directly passing this list into the UI, this avoids the issue of the UI not re-rendering.
- We can also simply make use of the add/set/delete operations under the UniqueCommissionList which makes things easier.
